refs #3157 - revert "fixes #10627 - apply host taxonomy scope to facts/report joins"
Rails 4 correctly applies default scopes even when doing complex joins,
so the `Report.joins(:host => [:hostgroup])` join would apply report,
host and hostgroup default scopes.

Both host and hostgroup have taxonomy-based default scopes which is
enough to resolve the original bug.

The extra change to the tests to set hostgroup taxonomies is because
the original fix only applied the host taxonomy scope and not hostgroup,
so this is now required to avoid mismatches.

This reverts commit 9926db4274e2d9d01466676116a9c001f473f1a4.